pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "QLever-UI"
version = "0.1.0"
description = "A user-friendly interface for the QLever SPARQL engine"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django==5.0.3",
    "django-environ==0.11.2",
    "django-filter>=25.1",
    "django-import-export==3.3.7",
    "djangorestframework>=3.16.1",
    "gunicorn==22.0.0",
    "markdown>=3.9",
    "requests==2.32.0",
    "whitenoise[brotli]==6.6.0",
]

[tool.setuptools.packages.find]
include = ["qlever*", "backend*", "resources*", "formatter*"]

[tool.ruff]
line-length = 79

[tool.ruff.lint]
extend-select = ["I"]
